+++ src/js/helper/specialEntry.js
@@ -127,14 +127,14 @@
                     let parentId = parentIds[i];
 
                     if (typeof parentId !== "undefined") {
-                        let parentEntry = ext.elements.bookmarkBox.all.find("a[data-id='" + parentId + "']");
+                        let parentEntry = ext.elements.bookmarkBox.all.find("a[" + ext.opts.attr.id + "='" + parentId + "']");
 
                         if (parentEntry.length() > 0 && parentEntry.next("ul").length() > 0) {
                             let index = 0;
                             let infos = [];
 
                             parentEntry.next("ul").children("li").forEach((entry) => {
-                                if (!$(entry).hasClass(ext.opts.classes.drag.dragInitial)) {
+                                if (!$(entry).hasClass(ext.opts.classes.drag.dragInitial) && $(entry).find("span.removeMask").length() === 0) {
                                     if ($(entry).children("a." + ext.opts.classes.sidebar.separator).length() > 0) {
                                         infos.push({index: index});
                                     } else {